fix auto verify
