+++ src/service/verifyService.ts
@@ -1,77 +1,18 @@
-import {
-    GuildMember,
-    Guild,
-    PartialGuildMember,
-    MessageEmbed,
-} from "discord.js";
+import { Collection, GuildMember, MessageEmbed, OAuth2Guild } from "discord.js";
 import { Guild as GuildEntity } from "../entity/Guild";
 
 import { Team } from "../service/util/constants";
-import { setNickname } from "../service/nickname";
+import { setNickname } from "./nicknameService";
 import { UserData } from "../wrapper/models/user";
-import { ApiError, BotError } from "../error";
+import { BotError } from "../error";
 import { AppDataSource } from "..";
-import { EventHandler, DiscordEvent } from "../decorators";
 import { makeLogger } from "../logger";
+import { Color } from "../service/util/constants";
 import { UserRank } from "../entity/UserRank";
-import { removeMembers } from "../service/nickname";
-import { isSome } from "../util/guard";
 import { getUser } from "../wrapper/wrapper";
 import { config } from "../config";
-import { Color } from "../service/util/constants";
 
 const logger = makeLogger(module);
-
-@EventHandler()
-export class VerifyService {
-    @DiscordEvent("guildMemberAdd")
-    async initUserRank(member: GuildMember) {
-        logger.info(`${member.displayName} joined ${member.guild.name}`);
-        try {
-            const userData = await getUser(config.apiKey, member.id);
-            const guild = await getGuild(member.guild.id);
-            updateRoleAndNickname(userData, guild, member);
-            sendMessage(
-                member,
-                `Successfully verified user ${member.user.tag}!`,
-                Color.GREEN
-            );
-        } catch (e) {
-            if (e instanceof ApiError && e.code === 2) {
-                sendMessage(
-                    member,
-                    `Your discord account is not verified with Final Earth.
-                Please visit [here](https://www.finalearth.com/discord) and follow the instructions.`,
-                    Color.RED
-                );
-            } else {
-                logger.error(e);
-            }
-        }
-    }
-
-    @DiscordEvent("guildMemberRemove")
-    async memberLeft(member: GuildMember | PartialGuildMember) {
-        logger.info(`${member.displayName} left ${member.guild.name}`);
-        const repository = AppDataSource.getRepository(UserRank);
-        const userRank = await removeMembers(member, repository);
-        if (userRank) repository.save(userRank);
-    }
-
-    @DiscordEvent("guildDelete")
-    async guildDelete(guild: Guild) {
-        logger.info(`Bot kicked from guild: ${guild.name}`);
-        const repository = AppDataSource.getRepository(UserRank);
-        const members = await guild.members.fetch();
-        const users = (
-            await Promise.all(
-                members.map((member) => removeMembers(member, repository))
-            )
-        ).filter(isSome);
-
-        if (users.length > 0) repository.save(users);
-    }
-}
 
 export async function updateRoleAndNickname(
     user: UserData,
@@ -121,7 +62,11 @@ export async function getGuild(guildId: string): Promise<GuildEntity> {
     });
 }
 
-async function sendMessage(member: GuildMember, message: string, color: Color) {
+export async function sendMessage(
+    member: GuildMember,
+    message: string,
+    color: Color
+) {
     const guild = await getGuild(member.guild.id);
     if (!guild) {
         throw new BotError("Guild not found, something is very wrong.");
@@ -145,4 +90,29 @@ async function getVerifyChannel(member: GuildMember, guild: GuildEntity) {
     return guild.verify_channel
         ? member.guild.channels.cache.get(guild.verify_channel)
         : undefined;
+}
+
+export async function processUser(
+    u: UserRank,
+    guilds: Collection<string, OAuth2Guild>
+): Promise<void> {
+    try {
+        const user = await getUser(config.apiKey, u.discord_id);
+        if (guilds) {
+            for (const g of u.guild_ids) {
+                const guild = await guilds.get(g)?.fetch();
+                if (guild) {
+                    const guildEntity = await getGuild(guild.id ?? "");
+                    const member = await guild.members.fetch(u.discord_id);
+                    try {
+                        await updateRoleAndNickname(user, guildEntity, member);
+                    } catch (e) {
+                        logger.debug(e);
+                    }
+                }
+            }
+        }
+    } catch (e) {
+        logger.debug(e);
+    }
 }
